wip
- permanent Ids
- SmartMessageManager
